Fixed some warnings

+++ src/Sample_Scene.cpp
@@ -27,8 +27,8 @@ void StarBangBang::Sample_Scene::Init()
 		double rads = (2 * (double)PI / 8) * i;
 		AEVec2& pos = obj->transform.position;
 
-		pos.x = cos(rads) * 200;
-		pos.y = sin(rads) * 200;
+		pos.x = static_cast<float>(cos(rads) * 200);
+		pos.y = static_cast<float>(sin(rads) * 200);
 
 		objectManager.AddImage(obj, image);
 		objectManager.AddCollider(obj, true);
